Add confirm dialog for deleting attraction

+++ client/src/features/attraction/manage/AttractionFormPage.tsx
@@ -127,6 +127,8 @@ export default function AttractionFormPage() {
             variant="contained"
             color="warning"
             onClick={() => {
+              if (!confirm('Are you sure you want to delete this attraction?'))
+                return;
               deleteAttraction(id)
                 .then(() => {
                   toast.success('Attraction deleted');
